+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Show popups in Wellness ToDo and Rings just once [#2071](https://github.com/rokwire/illinois-app/issues/2071).
 
 ## [4.1.0] - 2022-07-20
 ### Changed